Update EnvironmentCreateForm.js to combine commands for policy creation and bootstrapping (#868)
### Feature or Bugfix
- Bugfix

### Detail
This PR is to combine cdk exec policy stack creation and cdk
bootstrapping for linking environments in the Link

### Relates
[- <URL or Ticket>](https://github.com/awslabs/aws-dataall/issues/688)

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized? Yes
- What precautions are you taking before deserializing the data you
consume? N/A
  - Is injection prevented by parametrizing queries? N/A
  - Have you ensured no `eval` or similar functions are used? N/A
- Does this PR introduce any functionality or component that requires
authorization? N/A
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
N/A
  - Are you logging failed auth attempts?N/A
- Are you using or adding any cryptographic features? N/A
  - Do you use a standard proven implementations? Yes
- Are the used keys controlled by the customer? Where are they stored?
N/A
- Are you introducing any new policies/roles/users? N/A
- Have you used the least-privilege principle? How? Yes. The CDK
execution role is restricted from Administrator privileges to
least-privilege principle with a custom policy.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.