Add additional checks for dataset importing (#883)
### Feature or Bugfix
Feature

### Detail
Added additional checks for dataset importing  (see #614 )

Implemented checks that fail if:
1) An importing dataset bucket is encrypted with AWS managed key, but
KMS key provided
2) An importing dataset bucket is encrypted with KMS key, but KMS key
not provided
3) An importing dataset bucket is encrypted with KMS key, but provided
KMS key is wrong
4) An imported dataset bucket is encrypted with AWS managed key, but
user is trying to create a share requests for different account than
dataset's account

From #614:
> User can forget to grant the pivotRole access to modify the key policy
-> All folder/bucket shares will fail when trying to update the key
policy

The check for this is not implemented. Rationale for this:
There are two types of pivotRole: manual pivotRole (1) and cdk pivotRole
(2).
1) Manual pivotRole has `kms:PutKeyPolicy` for '*' resources in the
environment. No action needed
2) When dataset stack is deployed (for imported datasets), it triggers
update of an environment stack. The environment stack updates policies
for all imported keys and datasets. Eventually, it sets
`kms:PutKeyPolicy` for the imported key that pivotRole can change the
policy. Thus, the users can catch the error only if they try to create
share before update of the stack is succeed

### Relates
#614

### Testing

Added unit test for checks in this PR. Added example of testing data.all
aws clients, by mocking boto3 client. Usually, aws clients are mocked in
the integration testing and this test can be used as an example how to
mock boto3 instead.

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Are you introducing any new policies/roles/users? Yes
- Adding `s3:GetEncryptionConfiguration` to the pivotRole to fetch the
encryption configuration for env buckets. It's needed to implement
additional checks. EncryptionConfiguration provides only information
about encryption (encryption type and key) and doesn't allow pivotRole
to read the content of the bucket.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: dlpzx <dlpzx@amazon.com>