Update CodeBuild images to Linux2 standard5.0 (node16 to node18) + Update Docker images to use AmazonLinux:2023 (node18 and Python3.9) (#889)
### Feature or Bugfix
- Bugfix

### Detail
The purpose of this PR is to upgrade any compute resource that uses
node16 to node18.

- CodeBuild images: [Amazon Linux 2 x86_64 standard:4.0 use node16
](https://docs.aws.amazon.com/codebuild/latest/userguide/available-runtimes.html)which
is already deprecated. In this PR we update the CodeBuild images to use
Amazon Linux 2 x86_64 standard:5.0 instead
- Docker images: In this PR we replace AmazonLinux2 images by
[AmazonLinux2023](https://docs.aws.amazon.com/linux/al2023/ug/what-is-amazon-linux.html),
the next generation of Amazon Linux from Amazon Web Services. In
AmazonLinux2023 the default Python version installed is 3.9. For this
reason we also upgrade the Python version in this PR.

### Relates
#782 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/). N/A

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.