Fix permissions on share workflows (#914)
### Feature or Bugfix
- Bugfix

### Detail
There has been a mismatch on the permissions-granting in the shares.
There are 2 types of permissions: on the share_object and preview
permissions for requesters in the shared_tabled

In this PR:
- delete share object permissions from RDS when the share object is
deleted
- attach table permissions to requesters for approved tables in
`approve_share_object`
- delete table permissions to requesters for revoked tables in
`revoke_items_share_object`
- delete table permissions to requesters for revoked tables in
`remove_shared_item` if the removed item is a table and was in
Share_Failed
- remove the deletion of permissions on the dataset level (which should
not have been there in the first place)

### Relates

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.