Update Worksheet database names in UI for new simplified db names (#1063)
### Feature or Bugfix
- Bugfix

### Detail
After implementing #1016 the names displayed for the databases in
Worksheets won't contain the unique identifier.
In addition this PR solves
https://github.com/data-dot-all/dataall/issues/805 by removing
duplicates also in FE.

Here is an screenshot of a local test:

![Screenshot 2024-02-14 at 17 34
16](https://github.com/data-dot-all/dataall/assets/71252798/a3756f7e-4284-4420-8d48-a33a6340120d)

Update:
Because there will be a mix of old shares with Glue database names
ending in`shared_URI` and shares with database names suffixed by
`shared` only, this PR introduces a new field in the GraphQL type
returned by the searchDataItems query. This field resolves the name of
the shared Glue database. @noah-paige @TejasRGitHub in this
[commit](https://github.com/data-dot-all/dataall/pull/1063/commits/3321109705976b66fe33fd9645c9e2bf5140b546)


At first I tried implementing a separate resolver for the Worksheets,
but I think we can fix step by step and first focus on the database name
and then on the group of the Worksheet vs the group chosen inside the
Worksheet. In any case I left the [commit
](https://github.com/data-dot-all/dataall/pull/1063/commits/82e4ed3f83f6138f5b605b9e20faa1b5e0034e72)to
have some reference.


### Relates
- #1016 
- #805 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.