Solving issue with worksheets for catalog Dbs (#1079)
### Feature or Bugfix
- Bugfix


### Detail

When using worksheet with a share made with a catalog account ( by using
steps as described here in this PR -
https://github.com/data-dot-all/dataall/pull/1021 ) , the worksheet drop
down list doesn't display the correct DB name. This is due to the fact
that DB name is picked from the producer account ( where the S3 bucket
is present and where the actualDB is not present ) which has the
resource linked DB. Thus, the autogenerated querying doesn't work .
Please refer to the screenshot
<img width="1482" alt="image"
src="https://github.com/data-dot-all/dataall/assets/71188245/fbc28286-0ca7-47de-a6ae-3020b1188dcb">

Also, on the share view, the db name mentioned on the query ( in the
"Data Consumption details" ) is the resource linked DB name and not the
correct DB name.

### Relates
- https://github.com/data-dot-all/dataall/issues/904

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)? No
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization? No
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features? No
  - Do you use a standard proven implementations?
- Are the used keys controlled by the customer? Where are they stored?
No
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: trajopadhye <tejas.rajopadhye@yahooinc.com>