Restricting Team/User Access to Worksheets (#1137)
### Feature or Bugfix

- Bugfix


### Detail
*IMPORTANT*
Environments stacks should be updated for the changes to be effective.

Glue-permissions were added to the team's env role only if the role had
CREATE_DATASET permissions. 
Now the required permissions are added to the team's env role regardless of CREATE_DATASET permission.
So, now it's possible to give the team permissions ONLY to query dataset via worksheet. 


### Relates
- [945](https://github.com/data-dot-all/dataall/issues/945)

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)? N/A
- Does this PR introduce any functionality or component that requires
authorization? N/A
- Are you using or adding any cryptographic features? N/A
- Are you introducing any new policies/roles/users? N/A


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Sofia Sazonova <sazonova@amazon.co.uk>