Fix missing GET_FOLDER permissions (#1163)
### Feature or Bugfix
- Bugfix

### Detail
- Adds a new type of RESOURCE permission: `GET_DATASET_FOLDER`
- check `GET_DATASET_FOLDER` against the folder when we execute the
`get_storage_location` service function
- attach `GET_DATASET_FOLDER` permission to the dataset owner and
stewards when the folder is created
- remove `GET_DATASET_FOLDER` permissions to the dataset owner and
stewards when the folder is deleted
- attach `GET_DATASET_FOLDER` permissions to the share requester group
when the folder is shared
- remove `GET_DATASET_FOLDER` permissions to the share requester group
when the folder is revoked
- Migration script that creates the new permission type and creates
records backfilling the existing datasets and shares corresponding
permissions.


NEXT STEPS: 
- This PR only grants TABLE and FOLDER READ permissions to share
requesters when they make a share as a GROUP. It does not grant
permissions to the group if the share request is done on behalf of a
CONSUMPTION ROLE. If we want to implement this feature, which we have to
decide, we can do it in a separate PR --> Issue
https://github.com/data-dot-all/dataall/issues/1174 and Issue
https://github.com/data-dot-all/dataall/issues/1173
- get_table service method is not decorated --> Issue
https://github.com/data-dot-all/dataall/issues/1171


Testing (locally)
- [X] Create folder and as dataset owner check folder details
- [X] Request access to folder and as requester check folder details
- [X] Revoke the access, the group cannot see the folder details anymore
- [X] Delete folder and Check database - permissions to folder are
deleted
- [X] Request access to folder with a consumption role - the requester
group does not have access to the folder


Backwards compatibility (checked in database as well)
- [X] After upgrading with migration script, the dataset owner can check
folder details
- [X] and the share request owner can check the folder details

<img width="995" alt="image"
src="https://github.com/data-dot-all/dataall/assets/71252798/2c1ee4c5-d0b3-4e67-bdd8-0701ef93d3aa">



### Relates
- #804 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.